Fix system hang by PMI
Generally, the system reacts well when placing some printk inside
an IRQ handler, but in particular cases, this leads to a system hang.
In our experimental setup, this happens when the system generates
a lot of threads (fork) - osbench: thread generations test.
To overcome this problem, all the prints inside the PMI handler have
been removed.

This commit implies minor fixes.

+++ driver/recode.h
@@ -46,7 +46,7 @@ extern unsigned __read_mostly fixed_pmc_pmi;
 extern unsigned __read_mostly max_pmc_fixed;
 extern unsigned __read_mostly max_pmc_general;
 
-extern unsigned __read_mostly max_pmi_before_ctx;
+extern unsigned long __read_mostly max_pmi_before_ctx;
 
 enum recode_state {
 	OFF = 0,
